+++ package.json
@@ -81,12 +81,12 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.4",
     "jest": "^29.7.0",
-    "nitrogen": "^0.30.2",
+    "nitrogen": "^0.31.5",
     "prettier": "^3.6.2",
     "react": "19.1.0",
     "react-native": "0.81.4",
     "react-native-builder-bob": "^0.40.13",
-    "react-native-nitro-modules": "^0.30.2",
+    "react-native-nitro-modules": "^0.31.5",
     "release-it": "^19.0.4",
     "turbo": "^2.5.6",
     "typescript": "^5.9.2"
@@ -94,7 +94,7 @@
   "peerDependencies": {
     "react": "*",
     "react-native": "*",
-    "react-native-nitro-modules": "^0.30.2"
+    "react-native-nitro-modules": "^0.31.5"
   },
   "workspaces": [
     "example"
